requirement.txt
accelerate==1.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
diffusers==0.33.1
filelock==3.13.1
fonttools==4.58.1
fsspec==2024.6.1
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.3
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
peft==0.15.2
pillow==11.0.0
protobuf==6.31.1
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.9.0
urllib3==2.4.0
zipp==3.22.0
